requirements.txt
dj-config-url
requests==2.18.4
Django==2.1.5
beautifulsoup4==4.7.1
whitenoise==4.1.2
pytz==2018.9
gunicorn==19.9.0
psycopg2==2.7.6.1
djangorestframework==3.9.0